requirements.txt
streamlit>=1.34
langgraph
pydantic>=2.6
openai>=1.50
python-dotenv>=1.0
pandas>=2.2
numpy>=1.26
tqdm>=4.66
python-docx>=1.1.2
